Update README to fix bad command. Refactor Fibonacci method to remove int cast as it disallows returning of very large numbers.

+++ app/Fibonacci.php
@@ -29,7 +29,7 @@ class Fibonacci {
         self::_validate($number);
 
         // Another existing algorithm thanks to slack.
-        return (int) round(((5 ** .5 + 1) / 2) ** ($number-1) / 5 ** .5);
+        return round(((5 ** .5 + 1) / 2) ** ($number-1) / 5 ** .5);
     }
 
     private static function _validate($number) {
